Addressing feedback

+++ code/graphics/light.cpp
@@ -138,7 +138,7 @@ void FSLight2GLLight(light* FSLight, gr_light* GLLight) {
 		break;
 
 	default:
-		Error(LOCATION, "Unknown light type in FSLight2GLLight. Expected was 0, 1, 2 or 3, we got %i", FSLight->type);
+		Error(LOCATION, "Unknown light type in FSLight2GLLight. Expected was 0, 1, 2 or 3, we got %i", static_cast<int>(FSLight->type));
 		break;
 	}
 }
